Final script to push messages to dynamo tables

+++ src/python/send_messages_sqs.py
@@ -45,6 +45,6 @@ if __name__ == '__main__':
         log.info(f"Message sent: {resp['MessageId']} ::> {message}")
         i += 1
         # wait
-        sleep(int(args.interval))
+        sleep(float(args.interval))
 
     log.info("Ending Program")
